+++ src/funcs/collectionsList.ts
@@ -38,6 +38,10 @@ export function collectionsList(
     | errors.ResourceNotFoundError
     | errors.TooManyRequestsError
     | errors.InternalServerError
+    | errors.PayloadTooLargeError
+    | errors.BadGatewayError
+    | errors.ServiceUnavailableError
+    | errors.GatewayTimeoutError
     | LambdaDBError
     | ResponseValidationError
     | ConnectionError
@@ -67,6 +71,10 @@ async function $do(
       | errors.ResourceNotFoundError
       | errors.TooManyRequestsError
       | errors.InternalServerError
+      | errors.PayloadTooLargeError
+      | errors.BadGatewayError
+      | errors.ServiceUnavailableError
+      | errors.GatewayTimeoutError
       | LambdaDBError
       | ResponseValidationError
       | ConnectionError
@@ -155,6 +163,10 @@ async function $do(
     | errors.ResourceNotFoundError
     | errors.TooManyRequestsError
     | errors.InternalServerError
+    | errors.PayloadTooLargeError
+    | errors.BadGatewayError
+    | errors.ServiceUnavailableError
+    | errors.GatewayTimeoutError
     | LambdaDBError
     | ResponseValidationError
     | ConnectionError